tests/wpt/tests/html/dom/render-blocking/parser-inserted-inline-module-with-import.html

<!DOCTYPE html>
<head>
<title>Parser-inserted module script elements with "blocking=render" are render-blocking</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  window.did_execute_script = false;
</script>
<script type="module" blocking="render">
  import "/loading/resources/dummy.js?pipe=trickle(d1)";
  window.did_execute_script = true;
</script>
</head>
<div id="dummy">some text</div>

<script>
    promise_test(async t => {
      await new Promise(resolve => requestAnimationFrame(() => resolve()));
      assert_true(window.did_execute_script, "Parser-inserted render-blocking inline module script should execute before rAF callback");
    });
</script>
